Cleaned up some classes and added the boilerplate for global services and an assetManager

+++ spirit_cards/core/entity.py
@@ -1,9 +1,9 @@
 class Entity:
 
-    def update(self, delta: int) -> None:
+    def update(self, delta: float) -> None:
         pass
 
-    def render(self, delta: int) -> None:
+    def render(self, delta: float) -> None:
         pass
 
     def cleanup(self) -> None:
